requirements.txt
boto3==1.17.100
botocore==1.20.100
certifi==2021.5.30
chardet==4.0.0
elasticsearch==7.13.1
idna==2.10
jmespath==0.10.0
python-dateutil==2.8.1
requests==2.25.1
requests-aws4auth==1.1.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.5